[#4027] Remove thrift dependency from pinpoint-commons
