+++ commons/pom.xml
@@ -12,19 +12,6 @@
     <packaging>jar</packaging>
 
     <dependencies>
-        <dependency>
-            <groupId>com.navercorp.pinpoint</groupId>
-            <artifactId>pinpoint-thrift</artifactId>
-            <scope>provided</scope>
-        </dependency>
-
-        <!-- commons module may be used at agent so there will be problems. you should set optional -->
-        <dependency>
-            <groupId>org.apache.thrift</groupId>
-            <artifactId>libthrift</artifactId>
-            <optional>true</optional>
-        </dependency>
-
 
         <!-- Logging dependencies -->
         <dependency>
